[build] Update compilation flags and fix all warnings

+++ src/MTracks.hpp
@@ -20,7 +20,7 @@ class MTracks : public QObject
 
     Q_PROPERTY(QUrl matchingFolder READ getMatchingFolder WRITE setMatchingFolder NOTIFY matchingFolderChanged)
     Q_PROPERTY(Status status READ status NOTIFY statusChanged)
-    Q_PROPERTY(int nbTracks READ nbTracks CONSTANT)
+    Q_PROPERTY(size_t nbTracks READ nbTracks CONSTANT)
 
 public:
     enum Status {
@@ -74,7 +74,6 @@ public:
 
     QUrl getMatchingFolder() const { return _matchingFolder; }
     void setMatchingFolder(const QUrl& matchingFolder) {
-
         if(matchingFolder == _matchingFolder)
             return;
        _matchingFolder = matchingFolder;
@@ -93,7 +92,7 @@ public:
        }
     }
 
-    inline int nbTracks() const {
+    inline size_t nbTracks() const {
         if(!_tracks || _status != MTracks::Ready)
             return 0;
         return _tracks->size();
@@ -104,7 +103,6 @@ private:
     std::unique_ptr<aliceVision::track::TracksMap> _tracks;
     std::unique_ptr<aliceVision::track::TracksPerView> _tracksPerView;
     Status _status = MTracks::None;
-
 };
 
 }
